added missoginie

+++ controllers/socialUrls/removeURL.js
@@ -8,11 +8,31 @@ const deleteURL = async (url, URLmodel, name) => {
     if (!element) {
       return { state: false, message: "This URL is not present" };
     }
-    doc = await URLmodel.findOneAndDelete({ url });
-    const { oAuth2Client } = await validateOauth(name);
-    const drive = google.drive({ version: "v3", auth: oAuth2Client });
-    await deleteDocsFromDrive([doc], drive);
-    return { state: true, message: "URL removed successfully" };
+
+    let shouldDelete = false;
+    if (element?.isImage || element?.isCarousel) {
+      if (!element.uploadedToFb && !element.uploadedToInstagram)
+        shouldDelete = true;
+    } else if (
+      !element.uploadedToFb &&
+      !element.uploadedToInstagram &&
+      !element.uploadedToYoutube
+    )
+      shouldDelete = true;
+
+    if (shouldDelete) {
+      doc = await URLmodel.findOneAndDelete({ url });
+      const { oAuth2Client } = await validateOauth(name);
+      const drive = google.drive({ version: "v3", auth: oAuth2Client });
+      await deleteDocsFromDrive([doc], drive);
+      return { state: true, message: "URL removed successfully" };
+    } else {
+      return {
+        state: false,
+        message:
+          "seems like url is uploaded in at least one social media, contact developer",
+      };
+    }
   } catch (error) {
     return {
       state: false,
